FontAwesome now hosted locally & upgraded to 6.0.0
FontAwesome recently upgraded to version 6 which caused some issues. The icons were no longer loading because of a CORS issue. I used this opportunity to download FontAwesome and use it locally since there's going to be a performance boost and it also quickly resolves the issue at hand. The icon names were slightly changed so I went through them all and changed them to match version 6. This also helped me find a bug where in the users page, the user's username would not show up on the navigation bar in the top right corner. These changes are reflected in the 'routes/users.js' file.